[refactor]: auth 관련 파일 패키지 구조 변경
