+++ cubco-api/src/main/java/org/cubco/auth/jwt/JWTutil.java
@@ -1,17 +1,13 @@
-package org.cubco.util;
+package org.cubco.auth.jwt;
 
 import io.jsonwebtoken.Jwts;
-import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 
 import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.nio.charset.StandardCharsets;
 import java.util.Date;
 
 @Component
